Freescale: ENGR00275483-5 ARM: imx6qdl-sabresd: enable mipi dsi and display
This patch enables the mipi dsi driver and TRULY-WVGA mipi lcd panel
driver on the imx6qdl-sabresd platforms. To control the panel's reset
and power enable functions, this patch adds two pins in the hoggrp node
as the gpio control pins and adds the mipi_dsi_pwr_on/mipi-dsi-reset
nodes for related drivers to use.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>