Freescale: ENGR00269945: ARM: imx6: add the secondary sabreauto dts for pin conflict
The patch is to solve the pin conflicts between devices that are
currently added in imx6qdl-sabreauto dts file.  It has ecspi1, i2c3, and
uart3 enabled while gpmi and weim disabled in the primary
imx6qdl-sabreauto.dtsi, and creates the secondary
imx6q/dl-sabreauto dts to have gpmi and weim enabled while others
disabled.  Since usbh1 and usbotg depend on GPIO from max7310 on i2c3,
they have to be disabled as well in the secondary sabreauto dts files.

Signed-off-by: Shawn Guo <shawn.guo@freescale.com>