Freescale: ENGR00278646-11 mmc: sdhci-esdhc-imx: improve sd3.0 stability
We observed on some sd3.0 cards(Toshiba SDHC U1) that it may require
to reset host controller before sending the next tuning command
or the tuning may fail and cause the card can not work on uhs mode.
The root cause why the card fails on tuning without reset is still unknow.
This could be treated as a workaround before finding root cause.

Signed-off-by: Dong Aisheng <b29396@freescale.com>