cpuidle: ladder governor requires !NO_HZ
This makes the build system only select the governor that is
considered best suited for the clock type. There are various
patches floating around to do the same thing.  I am doing this
quick and dirty to avoid confusion of the cpu_idle governor
that is installed and used in dmesg.