Freescale: ENGR00289381 mxc: asrc: Fix RCU stall in output_task_worker
In kernel 3.10.17, RCU has its threshold to detect RCU stall. So it might be
risky for us to use a whole second to wait for the completion, which would
be surely returned within 100ms. Thus, we shrink the wait period so as to
circumvent some potential RCU stall issue.

This patch also moved pair_hold into spin_lock protection due to a race with
pair_hold in close() and release().

Signed-off-by: Nicolin Chen <b42378@freescale.com>