leds: add DT support for LEDs wired to supply and default brightness
Some PWM outputs are wired such that the LED they're controlling is
connected to supply rather than ground.  Therefore, the duty cycle
needs to be inverted to make the LED behave as it should do.

We also provide a way to specify the default brightness when a
trigger is not specified.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>