rtc: pcf8523: Auto-detect the proper capacitance
Instead of forcing the capacitance to high for all hardware, or
adding another setting to device-tree we auto-detect the proper
capacitance for the oscillator.  First we check the oscillator
for stability based on its power on state, if that fails
switch to the other capacitance and check for stability. The
oscillator can take up to 2 seconds to stabilize.

Signed-off-by: Jon Nettleton <jon.nettleton@gmail.com>