ARM: cubox-i: fix device-tree to support both fbdev and kms
This adds the necessary nodes so either the new KMS driver or
the legacy FBDEV driver can be used on the Cubox-i/Hummingboard.
There should ideally only be a single device-tree configuration
that supports both drivers as the device-tree is supposed to
describe the hardware not bend to the software.