USB OTG ID pin selection is a property of the carrier
USB OTG ID pin selection, and weather it's pulled up or down to
force host selection (like in CuBox-i) is a carrier property
and not microsom.

In theory a carrier that wants to use USB OTG ID can do that
by connecting GPIO_1 pin to the phyiscal ID pin of a micro USB connector.

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>