net: fec: Fix UDP receive throughput
When we are letting NAPI poll for data take a breath once we
have reached our budget.  This seems to be enough time to let
the system catch up and clear out some room to handle more
packets.

The ksoftirqd not completely starving userspace this allows
multi-core machines to get full UDP bandwidth.  The iMX6QS
only receives a meager boost of maybe 4-5Mbps.

This is only a workaround until the problem can be looked into
a bit more intensely.