drm: galcore: Add GEM support to the driver
This migrates over the Armada GEM support to the galcore driver.
I will be using the same userspace so we can keep most everything
the same.  It will probably be simpler because we probably don't
need the *dumb* functions.  I have not removed them though because
they may be useful for something I haven't thought of yet.