mxc: gpu_viv: Add better v4 support for dmabuf handling
the v4 driver now tracks MMU allocations on a per core basis.
This patch allows userspace to now pass along the hardwareType
the dmabuf memory region will be used for so it will be mapped
properly.  The v2 protocol is now supported by defaulting the
core it is mapped to as the 2D core because as of now that is
the only user of the v2 userspace.