Support for gpio connected ir senders/receivers
Only the HummingBoard i2eX has an onboard IR receiver.
Not all HummingBoard users are willing to solder one
on the other HummingBoard types. So, since
all HummingBoards also have GPIO capabilities, an IR
receiver connected to the GPIO should be able to work
as well. Also, none of the HummingBoards have an IR
sender, so having the possibility to support IR
senders in Lirc is welcome.

To add support for all these functions, the
Raspberry Pi lirc_rpi module was ported to the
HummingBoard (and renamed to lirc_hb).

The new module uses the sysfs gpio kernel functions
to interact with the gpio. The module should
therefor work fine on all i.MX6 devices sharing the
same GPIO pin numbering.