dts: imx6qdl: fixup pcie resources
The reference clocks used were pointing to lvds_gate previously
but a patch by freescale has made that a null clock, and lvds1_out
is not the proper clock to use.

Additionally the way the interrupts are setup in the driver we
take the main interrupt A for the PCIe controller and then
derive MSI from it by subtracting three.  Therefore the device
tree needs to pass the main interrupt 155 to the driver so
non MSI mode will work properly.