usb: chipidea: Allow handling the clock for an USB phy/hub for udoo board
First patch done by Fabio Estevam, the second is unknown - taken from mtx512
repository:
  https://github.com/mtx512/linux-imx/commits/imx_3.10.17_1.0.0_beta-udoo

for file ci_hdrc_imx.c, ci-hdrc-imx.txt
  1-2-chipidea-ci_hdrc_imx-Allow-handling-the-clock-for-an-USB-phy-hub.patch
for file usbmisc_imx.c
	unknown patch (found in mtx512 repository)

Original message from Fabio:
When using external USB PHY or USB hub, it is common that they require a clock
input.

Add a 'clk_phy' clock, so that it can be retrieved from the device tree and
enabled in the driver, so that the clock can properly drive the external
USB phy/hub.

Tested on a imx6q-udoo board, that connects via USBH1 to a USB2514 hub.

In this board the USB2514 is clocked from a 24MHz clock that comes from the
imx6q CLKO2 pin.